fisher-random-chess/backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.10.9
numpy==1.21.2
python-chess==1.9.0